requirements.txt
anthropic>=0.40.0
openai>=1.50.0
fastapi>=0.115.0
uvicorn>=0.30.0
schedule>=1.2.0
python-dotenv>=1.0.0
pandas>=2.1.0
jinja2>=3.1.0
pyyaml>=6.0
pydantic>=2.0.0
